# openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  version: 1.0.0
  description: |
    Трекер задач. Задачи доступны только владельцу: каждый запрос к ним требует access-токен.
    Все ошибки приходят в одном формате, см. схему Error.
servers:
  - url: http://localhost:8080
tags:
  - name: auth
  - name: tasks
  - name: notifications
  - name: health
paths:
  /healthz:
    get:
      tags: [health]
      summary: Процесс жив
      operationId: healthz
      responses:
        "200":
          description: Сервис отвечает
  /readyz:
    get:
      tags: [health]
      summary: Сервис готов принимать запросы
      operationId: readyz
      responses:
        "200":
          description: База и Redis доступны
        "503":
          description: Зависимость недоступна или сервис останавливается
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Регистрация
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "409":
          description: Email уже занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Вход
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tokens"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          description: Неверный email или пароль. Ответ одинаков в обоих случаях.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Обменять refresh-токен на новую пару
      description: Старый refresh-токен после обмена перестаёт работать.
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tokens"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Отозвать refresh-токен
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Токен отозван
        "400":
          $ref: "#/components/responses/InvalidArgument"
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: Список задач пользователя
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Сколько задач вернуть
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Сколько задач пропустить. На больших списках листай через after_id, он не замедляется с глубиной.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: after_id
          in: query
          description: Вернуть задачи с id больше этого. Значение берётся из next_after_id предыдущей страницы.
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: done
          in: query
          description: Оставить только выполненные (true) или только невыполненные (false)
          schema:
            type: boolean
      responses:
        "200":
          description: Страница задач в порядке создания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [tasks]
      summary: Создать задачу
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTask"
      responses:
        "201":
          description: Задача создана
          headers:
            Location:
              description: Адрес созданной задачи
              schema:
                type: string
                example: /api/v1/tasks/1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/TooLarge"
  /api/v1/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      tags: [tasks]
      summary: Получить задачу
      operationId: getTask
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Задача
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [tasks]
      summary: Изменить часть полей задачи
      operationId: updateTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Задача после изменения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/TooLarge"
    delete:
      tags: [tasks]
      summary: Удалить задачу
      operationId: deleteTask
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Задача удалена
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notifications:
    get:
      tags: [notifications]
      summary: Уведомления пользователя
      description: |
        Уведомления хранит отдельный сервис notifier: этот запрос идёт к нему по gRPC.
        Уведомление появляется не мгновенно — событие о задаче проходит через Kafka.
        Если notifier не отвечает, запрос возвращает 503, остальное API при этом работает.
      operationId: listNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Сколько уведомлений вернуть
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: before_id
          in: query
          description: Вернуть уведомления старше этого. Значение берётся из next_before_id предыдущей страницы.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "200":
          description: Страница уведомлений, сначала новые
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/NotificationsUnavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из ответа /api/v1/auth/login. Живёт 15 минут.
  schemas:
    Credentials:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 72
    User:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    Tokens:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
          description: JWT для заголовка Authorization
        refresh_token:
          type: string
          description: Одноразовый токен для получения новой пары. Хранится на сервере, его можно отозвать.
        expires_in:
          type: integer
          description: Время жизни access-токена в секундах
          example: 900
    RefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token:
          type: string
    Task:
      type: object
      required: [id, title, done, created_at]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: купить молоко
        done:
          type: boolean
        due_at:
          type: string
          format: date-time
          description: Срок. Когда он наступит, воркер отправит напоминание.
        reminded_at:
          type: string
          format: date-time
          description: Когда отправлено напоминание
        created_at:
          type: string
          format: date-time
    NewTask:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        due_at:
          type: string
          format: date-time
    TaskPatch:
      type: object
      description: Нужно передать хотя бы одно поле.
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        done:
          type: boolean
        due_at:
          type: string
          format: date-time
          description: Новый срок сбрасывает отметку о напоминании.
    TaskList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
          description: Сколько задач подходит под фильтр
        limit:
          type: integer
        offset:
          type: integer
        next_after_id:
          type: integer
          format: int64
          description: Курсор следующей страницы. Отсутствует, если страница неполная.
    Notification:
      type: object
      required: [id, kind, task_id, title, created_at]
      properties:
        id:
          type: integer
          format: int64
        kind:
          type: string
          enum: [task.created, task.due]
          description: task.created — задача создана, task.due — срок задачи наступил.
        task_id:
          type: integer
          format: int64
        title:
          type: string
          description: Заголовок задачи на момент события
        created_at:
          type: string
          format: date-time
          description: Когда произошло событие
    NotificationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        next_before_id:
          type: integer
          format: int64
          description: Курсор следующей страницы. Отсутствует, если страница неполная.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - invalid_argument
                - unauthenticated
                - invalid_credentials
                - invalid_token
                - not_found
                - method_not_allowed
                - conflict
                - too_large
                - rate_limited
                - internal
            message:
              type: string
            fields:
              type: array
              description: Поля, не прошедшие проверку. Приходят все сразу.
              items:
                type: object
                required: [field, rule]
                properties:
                  field:
                    type: string
                    example: email
                  rule:
                    type: string
                    example: required
            request_id:
              type: string
              description: Идентификатор запроса. По нему запрос находится в логах сервера.
  responses:
    InvalidArgument:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: invalid_argument
              message: validation failed
              fields:
                - field: title
                  rule: required
              request_id: 01999f3a-7c11-7a52-9d3e-3b0f6a2c8e41
    Unauthorized:
      description: Токена нет, он неверный или просрочен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Задачи с таким id у пользователя нет. Чужая задача тоже отвечает 404.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooLarge:
      description: Тело запроса больше 1 МБ
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Слишком много попыток с одного адреса
      headers:
        Retry-After:
          description: Через сколько секунд повторить
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotificationsUnavailable:
      description: Сервис уведомлений не отвечает. Запрос можно повторить позже.
      headers:
        Retry-After:
          description: Через сколько секунд повторить
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: notifications_unavailable
              message: notifications are temporarily unavailable
              request_id: 01999f3a-7c11-7a52-9d3e-3b0f6a2c8e41